Cambios mayores en formulario y listado de datos del usuario

+++ app/src/main/java/com/example/registrocoches/Model/Coche.kt
@@ -20,5 +20,5 @@ data class Coche(
     val matricula: String,
     val marca: String,
     val caballos: Int,
-    val persona_id: Int
+    val persona_id: Long
 )
